Add basic login functionality

+++ digits/views.py
@@ -11,18 +11,16 @@ import werkzeug.exceptions
 from flask.ext.socketio import join_room, leave_room
 
 import digits
-from . import dataset, model
+from digits import dataset, model, utils
 from config import config_value
-from webapp import app, socketio, scheduler, autodoc
+from webapp import app, socketio, scheduler
 import dataset.views
 import model.views
-from digits.utils import errors
 from digits.utils.routing import request_wants_json
 from digits.log import logger
 
 @app.route('/index.json', methods=['GET'])
 @app.route('/', methods=['GET'])
-@autodoc(['home', 'api'])
 def home():
     """
     DIGITS home page
@@ -100,10 +98,51 @@ def get_job_list(cls, running):
             )
 
 
+### Authentication/login
+
+@app.route('/login', methods=['GET','POST'])
+def login():
+    """
+    Ask for a username (no password required)
+    Sets a cookie
+    """
+    # Get the URL to redirect to after logging in
+    next_url = utils.routing.get_request_arg('next') or \
+            flask.request.referrer or flask.url_for('home')
+
+    if flask.request.method == 'GET':
+        return flask.render_template('login.html', next=next_url)
+
+    # Validate username
+    username = utils.routing.get_request_arg('username').strip()
+    try:
+        utils.auth.validate_username(username)
+    except ValueError as e:
+        # Invalid username
+        flask.flash(e.message, 'danger')
+        return flask.render_template('login.html', next=next_url)
+
+    # Valid username
+    response = flask.make_response(flask.redirect(next_url))
+    response.set_cookie('username', username)
+    return response
+
+@app.route('/logout', methods=['GET','POST'])
+def logout():
+    """
+    Unset the username cookie
+    """
+    next_url = utils.routing.get_request_arg('next') or \
+            flask.request.referrer or flask.url_for('home')
+
+    response = flask.make_response(flask.redirect(next_url))
+    response.set_cookie('username', '', expires=0)
+    return response
+
+
 ### Jobs routes
 
 @app.route('/jobs/<job_id>', methods=['GET'])
-@autodoc('jobs')
 def show_job(job_id):
     """
     Redirects to the appropriate /datasets/ or /models/ page
@@ -120,7 +159,7 @@ def show_job(job_id):
         raise werkzeug.exceptions.BadRequest('Invalid job type')
 
 @app.route('/jobs/<job_id>', methods=['PUT'])
-@autodoc('jobs')
+@utils.auth.requires_login(redirect=False)
 def edit_job(job_id):
     """
     Edit a job's name and/or notes
@@ -128,6 +167,9 @@ def edit_job(job_id):
     job = scheduler.get_job(job_id)
     if job is None:
         raise werkzeug.exceptions.NotFound('Job not found')
+
+    if not utils.auth.has_permission(job, 'edit'):
+        raise werkzeug.exceptions.Forbidden()
 
     # Edit name
     if 'job_name' in flask.request.form:
@@ -150,7 +192,6 @@ def edit_job(job_id):
 @app.route('/datasets/<job_id>/status', methods=['GET'])
 @app.route('/models/<job_id>/status', methods=['GET'])
 @app.route('/jobs/<job_id>/status', methods=['GET'])
-@autodoc('jobs')
 def job_status(job_id):
     """
     Returns a JSON objecting representing the status of a job
@@ -169,7 +210,7 @@ def job_status(job_id):
 @app.route('/datasets/<job_id>', methods=['DELETE'])
 @app.route('/models/<job_id>', methods=['DELETE'])
 @app.route('/jobs/<job_id>', methods=['DELETE'])
-@autodoc('jobs')
+@utils.auth.requires_login(redirect=False)
 def delete_job(job_id):
     """
     Deletes a job
@@ -178,18 +219,21 @@ def delete_job(job_id):
     if job is None:
         raise werkzeug.exceptions.NotFound('Job not found')
 
+    if not utils.auth.has_permission(job, 'delete'):
+        raise werkzeug.exceptions.Forbidden()
+
     try:
         if scheduler.delete_job(job_id):
             return 'Job deleted.'
         else:
             raise werkzeug.exceptions.Forbidden('Job not deleted')
-    except errors.DeleteError as e:
+    except utils.errors.DeleteError as e:
         raise werkzeug.exceptions.Forbidden(str(e))
 
 @app.route('/datasets/<job_id>/abort', methods=['POST'])
 @app.route('/models/<job_id>/abort', methods=['POST'])
 @app.route('/jobs/<job_id>/abort', methods=['POST'])
-@autodoc('jobs')
+@utils.auth.requires_login(redirect=False)
 def abort_job(job_id):
     """
     Aborts a running job
@@ -204,7 +248,7 @@ def abort_job(job_id):
         raise werkzeug.exceptions.Forbidden('Job not aborted')
 
 @app.route('/clone/<clone>', methods=['POST', 'GET'])
-@autodoc('jobs')
+@utils.auth.requires_login
 def clone_job(clone):
     """
     Clones a job with the id <clone>, populating the creation page with data saved in <clone>
@@ -272,7 +316,6 @@ for code in HTTP_STATUS_CODES:
 ### File serving
 
 @app.route('/files/<path:path>', methods=['GET'])
-@autodoc('util')
 def serve_file(path):
     """
     Return a file in the jobs directory
@@ -286,7 +329,6 @@ def serve_file(path):
 ### Path Completion
 
 @app.route('/autocomplete/path', methods=['GET'])
-@autodoc('util')
 def path_autocomplete():
     """
     Return a list of paths matching the specified preamble
